Begin Moving SOAP Interfaces into Service Classes
Handling of direct SOAP interaction, i.e. interaction with the
ChannelAdvisor API at it lowest level, is abstracted into a module with
classes that directly correspond to the service and method names
provided by the ChannelAdvisor API. An additional layer of abstraction
is provided for ease of use but the lower level will still be publish
its methods for interacting with ChannelAdvisor in a more direct way.