Refactor Specs to Validate SOAP Request Generation
Change the focus of specs in the Service classes to validating that the
XML is well formed and with the appropriate elements rather than
validating at the integration level. Before we were specifying in a way
that was partially validating how ChannelAdvisor's API behaved but now
the specs ensure that our code generates XML that is compatible with the
current version (v6) of the ChannelAdvisor API. Integration tests will
follow separately to validate that end-to-end web services calls are
successful.